Rename utils -> openfermion.
